Removing upgrade checks
This was originally put into place to stop upgrades early on if the source version was an unsupported upgrade. A problem with this design was exposed VBR 10a as it was released as a full ISO vs a patch file.

A full ISO uses the upgrade process to apply updates so we can no longer assume that upgrades will occur from the previous version of the product.

To resolve this behavior, I'm removing the source version verification checks. This should not be an issue as the software upgrade itself has these checks built-in.

Signed-off-by: Chris Arceneaux <carcenea@gmail.com>