Adding v11 default variables
This is the initial commit of these variables.

TBD: Variables still reflect v10 and need to be updated to v11.

Roadblock: I won't be able to fully update these files until v11 GA. All this means is this collection will be unable to download the v11 ISO prior to GA.
Signed-off-by: Chris Arceneaux <carcenea@gmail.com>